added profile settings dialog
